RDM-2055 Introduce fieldsPurger to follow good SRP principles.
