+++ src/app/shared/case-editor/case-edit-submit.html
@@ -34,7 +34,6 @@
                     href="javascript:void(0)">Change</a>
                 </td>
               </tr>
-              </ng-template>
             </ng-container>
           </ng-container>
         </ng-container>
