+++ src/app/shared/case-editor/case-edit.component.ts
@@ -9,6 +9,7 @@ import { WizardPage } from '../domain/wizard-page.model';
 import { FieldsUtils } from '../utils/fields.utils';
 import { ActivatedRoute, Router } from '@angular/router';
 import { ShowCondition } from '../conditional-show/conditional-show.model';
+import { FieldsPurger } from '../utils/fields.purger';
 
 @Component({
   selector: 'ccd-case-edit',
@@ -43,6 +44,7 @@ export class CaseEditComponent implements OnInit {
     private router: Router,
     private route: ActivatedRoute,
     private fieldsUtils: FieldsUtils,
+    private fieldsPurger: FieldsPurger,
   ) {}
 
   ngOnInit(): void {
@@ -58,19 +60,12 @@ export class CaseEditComponent implements OnInit {
     });
   }
 
-  buildCanShowPredicate(): Predicate<WizardPage> {
-    let currentState = this.fieldsUtils.mergeCaseFieldsAndFormFields(this.eventTrigger.case_fields, this.form.controls['data'].value);
-    return (page: WizardPage): boolean => {
-      return page.parsedShowCondition.match(currentState);
-    };
-  }
-
   getPage(pageId: string): WizardPage {
-    return this.wizard.getPage(pageId, this.buildCanShowPredicate());
+    return this.wizard.getPage(pageId, this.fieldsUtils.buildCanShowPredicate(this.eventTrigger, this.form));
   }
 
   first(): Promise<boolean> {
-    let firstPage = this.wizard.firstPage(this.buildCanShowPredicate());
+    let firstPage = this.wizard.firstPage(this.fieldsUtils.buildCanShowPredicate(this.eventTrigger, this.form));
     return this.router.navigate([firstPage ? firstPage.id : 'submit'], { relativeTo: this.route });
   }
 
@@ -80,16 +75,16 @@ export class CaseEditComponent implements OnInit {
   }
 
   next(currentPageId: string): Promise<boolean> {
-    this.clearHiddenFields(currentPageId);
+    this.fieldsPurger.clearHiddenFields(this.form, this.wizard, this.eventTrigger, currentPageId);
 
-    let nextPage = this.wizard.nextPage(currentPageId, this.buildCanShowPredicate());
+    let nextPage = this.wizard.nextPage(currentPageId, this.fieldsUtils.buildCanShowPredicate(this.eventTrigger, this.form));
     return this.router.navigate([nextPage ? nextPage.id : 'submit'], { relativeTo: this.route });
   }
 
   previous(currentPageId: string): Promise<boolean> {
-    this.clearHiddenFields(currentPageId);
+    this.fieldsPurger.clearHiddenFields(this.form, this.wizard, this.eventTrigger, currentPageId);
 
-    let previousPage = this.wizard.previousPage(currentPageId, this.buildCanShowPredicate());
+    let previousPage = this.wizard.previousPage(currentPageId, this.fieldsUtils.buildCanShowPredicate(this.eventTrigger, this.form));
     if (!previousPage) {
       return Promise.resolve(false);
     }
@@ -98,7 +93,7 @@ export class CaseEditComponent implements OnInit {
   }
 
   hasPrevious(currentPageId: string): boolean {
-    return this.wizard.hasPreviousPage(currentPageId, this.buildCanShowPredicate());
+    return this.wizard.hasPreviousPage(currentPageId, this.fieldsUtils.buildCanShowPredicate(this.eventTrigger, this.form));
   }
 
   cancel(): void {
@@ -109,82 +104,5 @@ export class CaseEditComponent implements OnInit {
     this.confirmation = confirmation;
     return this.router.navigate(['confirm'], {relativeTo: this.route});
   }
-
-  private clearHiddenFields(currentPageId) {
-    let formFields = this.form.getRawValue();
-    this.clearHiddenFieldForFieldShowCondition(currentPageId, formFields);
-    this.clearHiddenFieldForPageShowCondition(formFields);
-
-  }
-
-  private clearHiddenFieldForPageShowCondition(formFields) {
-    this.wizard.pages.forEach(wp => {
-      if (this.hasShowConditionPage(wp, formFields)) {
-          let condition = new ShowCondition(wp.show_condition);
-          if (this.isHidden(condition, formFields)) {
-            this.resetPage(wp);
-          }
-      }
-    });
-  }
-
-  private clearHiddenFieldForFieldShowCondition(currentPageId, formFields) {
-    let currentPage: WizardPage = this.wizard.getPage(currentPageId, this.buildCanShowPredicate());
-    currentPage.wizard_page_fields.forEach(wpf => {
-      let case_field = this.findCaseFieldByWizardPageFieldId(currentPage, wpf);
-      if (this.hasShowConditionField(case_field, formFields)) {
-          let condition = new ShowCondition(case_field.show_condition);
-          if (this.isHidden(condition, formFields)) {
-            this.resetField(case_field);
-          }
-      }
-    });
-  }
-
-  private isHidden(condition, formFields) {
-    return !condition.match(formFields.data);
-  }
-
-  private findCaseFieldByWizardPageFieldId(currentPage, wizardPageField) {
-    return currentPage.case_fields.find(cf => cf.id === wizardPageField.case_field_id);
-  }
-
-  private hasShowConditionPage(wizardPage, formFields): boolean {
-    return wizardPage.show_condition && formFields.data[this.getShowConditionKey(wizardPage.show_condition)];
-  }
-
-  private hasShowConditionField(case_field, formFields): boolean {
-    return case_field.show_condition && formFields.data[this.getShowConditionKey(case_field.show_condition)];
-  }
-
-  private getShowConditionKey(show_condition) {
-    return show_condition.split('=')[0];
-  }
-
-  private resetField(field) {
-    if (Array.isArray(field.value)) {
-      field.value.splice(0, field.value.length);
-    } else if (this.isObject(field.value)) {
-      field.value = {};
-    } else {
-      field.value = '';
-    }
-    (this.form.get('data') as FormGroup).removeControl(field.id);
-  }
-
-  private resetPage(wizardPage: WizardPage) {
-    wizardPage.wizard_page_fields.forEach(wpf => {
-      let case_field = this.findCaseFieldByWizardPageFieldId(wizardPage, wpf);
-      this.resetField(case_field);
-    });
-  }
-
-  private getType(elem): string {
-    return Object.prototype.toString.call(elem).slice(8, -1);
-  }
-
-  private isObject(elem) {
-    return this.getType(elem) === 'Object';
-  };
 
 }
